flask-dashboard/requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Limiter==2.9.0
SQLAlchemy==2.0.19
mysql-connector-python==8.1.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.2.3
bcrypt==4.0.1
gunicorn==21.2.0
redis==4.5.5
